+++ client/src/screens/hub/HubExperimentsDashboardScreen/HubExperimentsDashboardScreen.js
@@ -214,9 +214,10 @@ class HubExperimentsDashboardScreen extends React.Component {
         });
 
         return {
-          runs: data.runs,
+          runs: data.runs.sort((a, b) => b.date - a.date),
           experiments: _.uniq(experiments),
           selectedExperiments: [],
+          selectedRuns: [],
           coloredCols: coloredCols
         };
       });