+++ client/src/screens/hub/HubMainScreen/components/ContextBox/ContextBox.js
@@ -12,6 +12,7 @@ import { buildUrl, classNames, sortOnKeys, formatValue, roundValue } from '../..
 import UI from '../../../../../ui';
 import { HUB_PROJECT_EXPERIMENT } from '../../../../../constants/screens';
 import ColumnGroupPopup from './components/ColumnGroupPopup/ColumnGroupPopup';
+import GroupConfigPopup from './components/GroupConfigPopup/GroupConfigPopup';
 
 class ContextBox extends Component {
   paramKeys = {};
@@ -456,49 +457,151 @@ class ContextBox extends Component {
                   ),
                   expandable: true
                 },
-                value: `min: ${min !== null && min !== undefined ? roundValue(min) : '-'} / avg: ${avg !== null && avg !== undefined ? roundValue(avg) : '-'} / max: ${max !== null && max !== undefined ? roundValue(max) : '-'}`,
+                context: {
+                  content: (
+                    <div className='ContextBox__table__item-aggregated_labels'>
+                      {
+                        !!traceModel.contexts?.length
+                          ? (
+                            <UI.Label className='ContextBox__table__item-aggregated_label' color={color}>
+                              {traceModel.contexts[0]}
+                            </UI.Label>
+                          )
+                          : '-'
+                      }
+                      {
+                        traceModel.contexts?.length > 1 && (
+                          <UI.Label
+                            className='ContextBox__table__item-aggregated_label'
+                            color={color}
+                            rounded
+                          >
+                            <UI.Tooltip tooltip={traceModel.contexts.slice(1).join(', ')}>
+                              +{traceModel.contexts.length - 1}
+                            </UI.Tooltip>
+                          </UI.Label>
+                        )
+                      }
+                    </div>
+                  ),
+                  expandable: true
+                },
+                value: {
+                  // content: (
+                  //   <UI.Label className='ContextBox__table__item-aggregated_label' color={color}>
+                  //     min: {min !== null && min !== undefined ? roundValue(min) : '-'} / avg: {avg !== null && avg !== undefined ? roundValue(avg) : '-'} / max: {max !== null && max !== undefined ? roundValue(max) : '-'}
+                  //   </UI.Label>
+                  // )
+                  content: (
+                    <div className='ContextBox__table__item-aggregated_labels'>
+                      <UI.Label
+                        className='ContextBox__table__item-aggregated_label'
+                        iconLeft={
+                          <UI.Tooltip tooltip={
+                            'Minimum value'
+                          }>
+                            <UI.Icon
+                              i='vertical_align_bottom'
+                              className='ContextBox__table__item-aggregated_icon'
+                              scale={1.2}
+                            />
+                          </UI.Tooltip>
+                        }
+                        rounded
+                        outline
+                      >
+                        {min !== null && min !== undefined ? roundValue(min) : '-'}
+                      </UI.Label>
+                      <UI.Label
+                        className='ContextBox__table__item-aggregated_label'
+                        iconLeft={
+                          <UI.Tooltip tooltip={
+                            'Average value'
+                          }>
+                            <UI.Icon
+                              i='vertical_align_center'
+                              className='ContextBox__table__item-aggregated_icon'
+                              scale={1.2}
+                            />
+                          </UI.Tooltip>
+                        }
+                        outline
+                        rounded
+                      >
+                        {avg !== null && avg !== undefined ? roundValue(avg) : '-'}
+                      </UI.Label>
+                      <UI.Label
+                        className='ContextBox__table__item-aggregated_label'
+                        iconLeft={
+                          <UI.Tooltip tooltip={
+                            'Maximum value'
+                          }>
+                            <UI.Icon
+                              i='vertical_align_top'
+                              className='ContextBox__table__item-aggregated_icon'
+                              scale={1.2}
+                            />
+                          </UI.Tooltip>
+                        }
+                        outline
+                        rounded
+                      >
+                        {max !== null && max !== undefined ? roundValue(max) : '-'}
+                      </UI.Label>
+                    </div>
+                  )
+                },
                 step: {
                   content: stepData !== null && stepData[1] !== null ? stepData[1] : '-',
                 },
               },
               config: (
                 <>
+                  <GroupConfigPopup
+                    config={traceModel.config}
+                  />
                   {
                     this.context.traceList?.grouping?.chart?.length > 0 && (
-                      <div className='ContextBox__table__group-indicator__chart'>
-                        <UI.Text small>{traceModel.chart + 1}</UI.Text>
-                      </div>
+                      <UI.Tooltip tooltip='Group Chart ID'>
+                        <div className='ContextBox__table__group-indicator__chart'>
+                          {traceModel.chart + 1}
+                        </div>
+                      </UI.Tooltip>
                     )
                   }
                   {
                     this.context.traceList?.grouping?.color?.length > 0 && (
-                      <div
-                        className='ContextBox__table__group-indicator__color'
-                        style={{
-                          backgroundColor: traceModel.color,
-                          borderColor: traceModel.color
-                        }}
-                      />
+                      <UI.Tooltip tooltip='Group Color'>
+                        <div
+                          className='ContextBox__table__group-indicator__color'
+                          style={{
+                            backgroundColor: traceModel.color,
+                            borderColor: traceModel.color
+                          }}
+                        />
+                      </UI.Tooltip>
                     )
                   }
                   {
                     this.context.traceList?.grouping?.stroke?.length > 0 && (
-                      <svg
-                        className='ContextBox__table__group-indicator__stroke'
-                        style={{
-                          borderColor: traceModel.color
-                        }}
-                      >
-                        <line
-                          x1='0'
-                          y1='50%'
-                          x2='100%'
-                          y2='50%'
+                      <UI.Tooltip tooltip='Group Stroke Style'>
+                        <svg
+                          className='ContextBox__table__group-indicator__stroke'
                           style={{
-                            strokeDasharray: traceModel.stroke.split(' ').map(elem => (elem / 5) * 3).join(' ')
+                            borderColor: traceModel.color
                           }}
-                        />
-                      </svg>
+                        >
+                          <line
+                            x1='0'
+                            y1='50%'
+                            x2='100%'
+                            y2='50%'
+                            style={{
+                              strokeDasharray: traceModel.stroke.split(' ').map(elem => (elem / 5) * 3).join(' ')
+                            }}
+                          />
+                        </svg>
+                      </UI.Tooltip>
                     )
                   }
                 </>