+++ client/src/ui/components/Table/Table.js
@@ -36,7 +36,10 @@ function Table(props) {
 
   return (
     <div className='Table__container'>
-      <div className='Table'>
+      <div className={classNames({
+        Table: true,
+        'Table--grouped': props.groups
+      })}>
         {
           leftPane && (
             <div className='Table__pane Table__pane--left'>
@@ -154,7 +157,7 @@ function Column({
       </div>
       {
         groups ? Object.keys(data).map(groupKey => (
-          <Fragment key={groupKey}>
+          <div key={groupKey} className='Table__group'>
             <div
               className={classNames({
                 Table__group__config__cell: true,
@@ -188,7 +191,8 @@ function Column({
               } : data[groupKey].data[col.key]}
               className={classNames({
                 Table__group__header__cell: true,
-                expanded: expanded[groupKey]
+                expanded: expanded[groupKey],
+                expandable: typeof data[groupKey].data[col.key] === 'object' && data[groupKey].data[col.key].expandable
               })}
             />
             {
@@ -196,7 +200,7 @@ function Column({
                 <Cell key={col.key + i} col={col} item={item[col.key]} />
               ))
             }
-          </Fragment>
+          </div>
         )) : data.map((item, i) => (
           <Cell key={col.key + i} col={col} item={item[col.key]} />
         ))